+++ IGC/WrapperLLVM/include/llvmWrapper/IR/DerivedTypes.h
@@ -45,7 +45,7 @@ namespace IGCLLVM
 #if LLVM_VERSION_MAJOR <= 10
         return llvm::cast<llvm::VectorType>(pType)->getBitWidth();
 #else
-        return pType->getPrimitiveSizeInBits().getFixedSize();
+        return (uint32_t)pType->getPrimitiveSizeInBits().getFixedSize();
 #endif
     }
 }
